fix bugs to make To Do skill work
