--- solutions/Virtual-Assistant/src/csharp/skills/ToDoSkill/Dialogs/AddToDo/Resources/AddToDoResponses.tt
@@ -1,108 +1,0 @@
-﻿<#@ assembly name="Newtonsoft.Json.dll" #>
-<#@ template debug="false" hostspecific="true" language="C#" #>
-<#@ output extension=".cs" #>
-<#
-  var className = System.IO.Path.GetFileNameWithoutExtension(Host.TemplateFile);
-  var namespaceName = System.Runtime.Remoting.Messaging.CallContext.LogicalGetData("NamespaceHint");;
-  string myFile = System.IO.File.ReadAllText(this.Host.ResolvePath(className + ".en.json"));
-  var json = Newtonsoft.Json.JsonConvert.DeserializeObject<System.Collections.Generic.Dictionary<string, dynamic>>(myFile);
-#>
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-using System;
-using System.Collections.Generic;
-using System.Globalization;
-using System.IO;
-using System.Linq;
-using System.Runtime.CompilerServices;
-using Microsoft.Bot.Solutions.Dialogs;
-using Newtonsoft.Json;
-
-namespace <#= namespaceName #>
-{
-    /// <summary>
-    /// Calendar bot responses class.
-    /// </summary>
-    public static class <#= className #>
-    {
-        private const string JsonFileName = "<#=className#>.*.json";
-
-        private static Dictionary<string, Dictionary<string, BotResponse>> jsonResponses;
-
-        // Generated code:
-        // This code runs in the text json:
-<# foreach (var propertyName in json)  { #>
-        public static BotResponse <#= propertyName.Key.Substring(0, 1).ToUpperInvariant() + propertyName.Key.Substring(1) #> => GetBotResponse();
-
-<# } #>
-        private static Dictionary<string, Dictionary<string, BotResponse>> JsonResponses
-        {
-            get
-            {
-                if (jsonResponses != null)
-                {
-                    return jsonResponses;
-                }
-
-                jsonResponses = new Dictionary<string, Dictionary<string, BotResponse>>();
-                var dir = Path.GetDirectoryName(typeof(<#= className #>).Assembly.Location);
-                var resDir = Path.Combine(dir, "Dialogs\\AddToDo\\Resources");
-
-                var jsonFiles = Directory.GetFiles(resDir, JsonFileName);
-                foreach (var file in jsonFiles)
-                {
-                    var jsonData = File.ReadAllText(file);
-                    var responses = JsonConvert.DeserializeObject<Dictionary<string, BotResponse>>(jsonData);
-                    var key = new FileInfo(file).Name.Split(".")[1].ToLower();
-
-                    jsonResponses.Add(key, responses);
-                }
-
-                return jsonResponses;
-            }
-        }
-
-        private static BotResponse GetBotResponse([CallerMemberName] string propertyName = null)
-        {
-            var locale = CultureInfo.CurrentUICulture.Name;
-            var theK = GetJsonResponseKeyForLocale(locale, propertyName);
-
-            // fall back to parent language
-            if (theK == null)
-            {
-                locale = CultureInfo.CurrentUICulture.Name.Split("-")[0].ToLower();
-                theK = GetJsonResponseKeyForLocale(locale, propertyName);
-
-                // fall back to en
-                if (theK == null)
-                {
-                    locale = "en";
-                    theK = GetJsonResponseKeyForLocale(locale, propertyName);
-                }
-            }
-
-            var botResponse = JsonResponses[locale][theK ?? throw new ArgumentNullException(nameof(propertyName))];
-            return JsonConvert.DeserializeObject<BotResponse>(JsonConvert.SerializeObject(botResponse));
-        }
-
-        private static string GetJsonResponseKeyForLocale(string locale, string propertyName)
-        {
-            try
-            {
-                if (JsonResponses.ContainsKey(locale))
-                {
-                    return JsonResponses[locale].ContainsKey(propertyName) ?
-                        JsonResponses[locale].Keys.FirstOrDefault(k => string.Compare(k, propertyName, StringComparison.CurrentCultureIgnoreCase) == 0) :
-                        null;
-                }
-
-                return null;
-            }
-            catch (KeyNotFoundException)
-            {
-                return null;
-            }
-        }
-    }
-}
