+++ solutions/Virtual-Assistant/src/csharp/skills/ToDoSkill/Dialogs/Shared/ToDoSkillDialog.cs
@@ -259,6 +259,94 @@ namespace ToDoSkill
             }
         }
 
+        public async Task<DialogTurnResult> CollectToDoTaskContent(WaterfallStepContext sc, CancellationToken cancellationToken = default(CancellationToken))
+        {
+            return await sc.BeginDialogAsync(Action.CollectToDoTaskContent);
+        }
+
+        public async Task<DialogTurnResult> AskToDoTaskContent(WaterfallStepContext sc, CancellationToken cancellationToken = default(CancellationToken))
+        {
+            var state = await this._accessor.GetAsync(sc.Context);
+            if (!string.IsNullOrEmpty(state.TaskContent))
+            {
+                return await sc.NextAsync();
+            }
+            else
+            {
+                var prompt = sc.Context.Activity.CreateReply(ToDoSharedResponses.AskToDoContentText);
+                return await sc.PromptAsync(Action.Prompt, new PromptOptions() { Prompt = prompt });
+            }
+        }
+
+        public async Task<DialogTurnResult> AfterAskToDoTaskContent(WaterfallStepContext sc, CancellationToken cancellationToken = default(CancellationToken))
+        {
+            try
+            {
+                var state = await _accessor.GetAsync(sc.Context);
+                if (string.IsNullOrEmpty(state.TaskContent))
+                {
+                    if (sc.Result != null)
+                    {
+                        sc.Context.Activity.Properties.TryGetValue("OriginText", out var toDoContent);
+                        state.TaskContent = toDoContent != null ? toDoContent.ToString() : sc.Context.Activity.Text;
+                        return await sc.EndDialogAsync(true);
+                    }
+                    else
+                    {
+                        return await sc.BeginDialogAsync(Action.CollectToDoTaskContent);
+                    }
+                }
+                else
+                {
+                    return await sc.EndDialogAsync(true);
+                }
+            }
+            catch
+            {
+                await HandleDialogExceptions(sc);
+                throw;
+            }
+        }
+
+        public async Task<DialogTurnResult> AddToDoTask(WaterfallStepContext sc, CancellationToken cancellationToken = default(CancellationToken))
+        {
+            try
+            {
+                var state = await _accessor.GetAsync(sc.Context);
+                if (string.IsNullOrEmpty(state.OneNotePageId))
+                {
+                    await sc.Context.SendActivityAsync(sc.Context.Activity.CreateReply(ToDoSharedResponses.SettingUpOneNoteMessage));
+                }
+
+                var service = await _serviceManager.Init(state.MsGraphToken, state.OneNotePageId);
+                var page = await service.GetDefaultToDoPage();
+                state.OneNotePageId = page.Id;
+                await service.AddToDo(state.TaskContent, page.ContentUrl);
+                var todosAndPageIdTuple = await service.GetToDos();
+                state.OneNotePageId = todosAndPageIdTuple.Item2;
+                state.AllTasks = todosAndPageIdTuple.Item1;
+                state.ShowToDoPageIndex = 0;
+                var rangeCount = Math.Min(state.PageSize, state.AllTasks.Count);
+                state.Tasks = state.AllTasks.GetRange(0, rangeCount);
+                var toDoListAttachment = ToAdaptiveCardAttachmentForOtherFlows(
+                    state.Tasks,
+                    state.AllTasks.Count,
+                    state.TaskContent,
+                    ToDoSharedResponses.AfterToDoTaskAdded,
+                    ToDoSharedResponses.ShowToDoTasks);
+
+                var toDoListReply = sc.Context.Activity.CreateReply();
+                toDoListReply.Attachments.Add(toDoListAttachment);
+                await sc.Context.SendActivityAsync(toDoListReply);
+                return await sc.EndDialogAsync(true);
+            }
+            catch
+            {
+                await HandleDialogExceptions(sc);
+                throw;
+            }
+        }
+
         // Validators
         private Task<bool> TokenResponseValidator(PromptValidatorContext<Activity> pc, CancellationToken cancellationToken)
         {
